Shader Optimization & Fix
1. Use half precision for vectors & colors, keep full precision for positions.

2. Fix the calculation in marschner specular.

Function will return a 0 value when "s.Layer" is 0. This causes some pixels to have pure black colors.

+++ Assets/Shaders/Shell/FurSpecular.hlsl
@@ -22,14 +22,15 @@ struct SurfaceOutputFur
 	half Kappa;
 };
 
-inline float square(float x) {
+inline half square(half x)
+{
 	return x * x;
 }
 
-float acosFast(float inX)
+half acosFast(half inX)
 {
-	float x = abs(inX);
-	float res = -0.156583f * x + (0.5 * PI);
+	half x = abs(inX);
+	half res = -0.156583f * x + (0.5 * PI);
 	res *= sqrt(1.0f - x);
 	return (inX >= 0) ? res : PI - res;
 }
@@ -37,48 +38,50 @@ float acosFast(float inX)
 #define SQRT2PI 2.50663
 
 //Gaussian Distribution for M term
-inline float Hair_G(float B, float Theta)
+inline half Hair_G(half B, half Theta)
 {
 	return exp(-0.5 * square(Theta) / (B*B)) / (SQRT2PI * B);
 }
 
 
-inline float3 SpecularFresnel(float3 F0, float vDotH) {
+inline half3 SpecularFresnel(half3 F0, half vDotH)
+{
 	return F0 + (1.0f - F0) * pow(1 - vDotH, 5);
 }
 
-inline float3 SpecularFresnelLayer(float3 F0, float vDotH, float layer) {
-	float3 fresnel = SpecularFresnel(F0,  vDotH);
+inline half3 SpecularFresnelLayer(half3 F0, half vDotH, half layer)
+{
+	half3 fresnel = SpecularFresnel(F0,  vDotH);
     return (fresnel * layer) / (1 + (layer-1) * fresnel);
 }
 
 // Yan, Ling-Qi, etc, "An efficient and practical near and far field fur reflectance model."
-float3 FurBSDFYan(SurfaceOutputFur s, float3 L, float3 V, float3 N, float Shadow, float Backlit, float Area)
+half3 FurBSDFYan(SurfaceOutputFur s, half3 L, float3 V, half3 N, half Shadow, half Backlit, half Area)
 {
-	float3 S = 0;
+	half3 S = 0;
 
-	const float VoL = dot(V, L);
-	const float SinThetaL = dot(N, L);
-	const float SinThetaV = dot(N, V);
-	float cosThetaL = sqrt(max(0, 1 - SinThetaL * SinThetaL));
-	float cosThetaV = sqrt(max(0, 1 - SinThetaV * SinThetaV));
-	float CosThetaD = sqrt((1 + cosThetaL * cosThetaV + SinThetaV * SinThetaL) / 2.0);
+	const half VoL = dot(V, L);
+	const half SinThetaL = dot(N, L);
+	const half SinThetaV = dot(N, V);
+	half cosThetaL = sqrt(max(0, 1 - SinThetaL * SinThetaL));
+	half cosThetaV = sqrt(max(0, 1 - SinThetaV * SinThetaV));
+	half CosThetaD = sqrt((1 + cosThetaL * cosThetaV + SinThetaV * SinThetaL) / 2.0);
 
-	const float3 Lp = L - SinThetaL * N;
-	const float3 Vp = V - SinThetaV * N;
-	const float CosPhi = dot(Lp, Vp) * rsqrt(dot(Lp, Lp) * dot(Vp, Vp) + 1e-4);
-	const float CosHalfPhi = sqrt(saturate(0.5 + 0.5 * CosPhi));
+	const half3 Lp = L - SinThetaL * N;
+	const half3 Vp = V - SinThetaV * N;
+	const half CosPhi = dot(Lp, Vp) * rsqrt(dot(Lp, Lp) * dot(Vp, Vp) + 1e-4);
+	const half CosHalfPhi = sqrt(saturate(0.5 + 0.5 * CosPhi));
 
-	float n_prime = 1.19 / CosThetaD + 0.36 * CosThetaD;
+	half n_prime = 1.19 / CosThetaD + 0.36 * CosThetaD;
 
-	float Shift = 0.0499f;
-	float Alpha[] =
+	half Shift = 0.0499f;
+	half Alpha[] =
 	{
 		-0.0998,//-Shift * 2,
 		0.0499f,// Shift,
 		0.1996  // Shift * 4
 	};
-	float B[] =
+	half B[] =
 	{
 		Area + square(s.Roughness),
 		Area + square(s.Roughness) / 2,
@@ -86,11 +89,11 @@ float3 FurBSDFYan(SurfaceOutputFur s, float3 L, float3 V, float3 N, float Shadow
 	};
 
 	//float F0 = square((1 - 1.55f) / (1 + 1.55f));
-	float F0 = 0.04652;//eta=1.55f
+	half F0 = 0.04652;//eta=1.55f
 
-	float3 Tp;
-	float Mp, Np, Fp, a, h, f;
-	float ThetaH = SinThetaL + SinThetaV;
+	half3 Tp;
+	half Mp, Np, Fp, a, h, f;
+	half ThetaH = SinThetaL + SinThetaV;
 	// R
 	Mp = Hair_G(B[0], ThetaH - Alpha[0]);
 	Np = 0.25 * CosHalfPhi;
@@ -103,9 +106,9 @@ float3 FurBSDFYan(SurfaceOutputFur s, float3 L, float3 V, float3 N, float Shadow
 	h = CosHalfPhi * (1 + a * (0.6 - 0.8 * CosPhi));
 	f = SpecularFresnelLayer(F0, CosThetaD * sqrt(saturate(1 - h * h)), s.Layer).x;
 	Fp = square(1 - f);
-	float sinGammaTSqr = square((h * a));
-	float sm = sqrt(saturate(square(s.Kappa)-sinGammaTSqr));
-	float sc = sqrt(1 - sinGammaTSqr) - sm;
+	half sinGammaTSqr = square((h * a));
+	half sm = sqrt(saturate(square(s.Kappa)-sinGammaTSqr));
+	half sc = sqrt(1 - sinGammaTSqr) - sm;
 	Tp = pow(s.Albedo, 0.5 * sc / CosThetaD) * pow(s.MedulaAbsorb*s.MedulaScatter, 0.5 * sm / CosThetaD);
 	Np = exp(-3.65 * CosPhi - 3.98);
 	S += (Mp * Np) * (Fp * Tp) * Backlit;
@@ -133,7 +136,7 @@ float3 FurBSDFYan(SurfaceOutputFur s, float3 L, float3 V, float3 N, float Shadow
 	S += (Mp * Np) * (f * Tp);
 
 	//TRTs
-	float phi = acosFast(CosPhi);
+	half phi = acosFast(CosPhi);
 	// hacking approximate pre-integrated Dtrts based on Cn
 	Np = 0.05f * cos(1.5*phi+1.7) + 0.18f;
 	// still assume h = sqrt(3)/2
@@ -145,27 +148,27 @@ float3 FurBSDFYan(SurfaceOutputFur s, float3 L, float3 V, float3 N, float Shadow
 	return S;
 }
 
-float3 FurDiffuseKajiya(SurfaceOutputFur s, float3 L, float3 V, float3 N, half Shadow, float Backlit, float Area)
+half3 FurDiffuseKajiya(SurfaceOutputFur s, half3 L, float3 V, half3 N, half Shadow, half Backlit, half Area)
 {
-	float3 S = 0;
-	float KajiyaDiffuse = 1 - abs(dot(N, L));
+	half3 S = 0;
+	half KajiyaDiffuse = 1 - abs(dot(N, L));
 
-	float3 FakeNormal = SafeNormalize(V - N * dot(V, N));
+	half3 FakeNormal = SafeNormalize(V - N * dot(V, N));
 	N = FakeNormal;
 
 	// Hack approximation for multiple scattering.
-	float Wrap = 1;
-	float NoL = saturate((dot(N, L) + Wrap) / square(1 + Wrap));
-	float DiffuseScatter = (1 / PI) * lerp(NoL, KajiyaDiffuse, 0.33);// *s.Metallic;
-	float Luma = Luminance(s.Albedo);
-	float3 ScatterTint = pow(s.Albedo / Luma, 1 - Shadow);
+	half Wrap = 1;
+	half NoL = saturate((dot(N, L) + Wrap) / square(1 + Wrap));
+	half DiffuseScatter = (1 / PI) * lerp(NoL, KajiyaDiffuse, 0.33);// *s.Metallic;
+	half Luma = Luminance(s.Albedo);
+	half3 ScatterTint = pow(s.Albedo / Luma, 1 - Shadow);
 	S = sqrt(s.Albedo) * DiffuseScatter * ScatterTint;
 	return S;
 }
 
-float3 FurBxDF(SurfaceOutputFur s, float3 N, half3 V, half3 L, float Shadow, float Backlit, float Area)
+half3 FurBxDF(SurfaceOutputFur s, half3 N, float3 V, half3 L, half Shadow, half Backlit, half Area)
 {
-	float3 S = float3(0, 0, 0);
+	half3 S = half3(0, 0, 0);
 
 	S = FurBSDFYan(s, L, V, N, Shadow, Backlit, Area);
 	S += FurDiffuseKajiya(s, L, V, N, Shadow, Backlit, Area);
